Create common interface for all augmentations
