+++ dataset.py
@@ -7,8 +7,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 import torch
 import torchvision.transforms
+from augmentations import *
 from pydicom import dcmread
-
 
 
 def train_test_split(dataset, test_size=0.2):
@@ -34,7 +34,8 @@ def train_test_split(dataset, test_size=0.2):
 
 class dcm_dataset(torch.utils.data.Dataset):
 
-    def __init__(self, directory, transforms=None):
+    def __init__(self, directory,
+                 transforms=torchvision.transforms.Compose([Downsample(), ImageTransform()])):
         """
         directory needs to contain the 'stage_2_train_labels.csv' file and the
         'stage_2_train_images' subdirectory
@@ -78,11 +79,13 @@ class dcm_dataset(torch.utils.data.Dataset):
             bounding_box[:, bb[1]:bb[1] + bb[3], bb[0]:bb[0] +
                          bb[2]] = torch.tensor(1.)
 
+        sample = img, (target, bounding_box)
+
         if self.transforms:
 
-            img, bounding_box = self.transforms(img, bounding_box)
+            sample = self.transforms(sample)
 
-        return img, (target, bounding_box)
+        return sample
 
     def display(self):
 
@@ -167,82 +170,3 @@ class dcm_dataset(torch.utils.data.Dataset):
 
         self.id = self.id[idx_to_keep]
         self.update()
-
-
-class Downsample:
-
-    def __init__(self):
-        self.pool = torch.nn.AvgPool2d(2)
-
-    def __call__(self, x, target):
-        x = self.pool(x.unsqueeze(0)).squeeze(0)
-        target = self.pool(target)
-        return x, target
-
-class Crop():
-    def __init__(self, max_shift):
-        self.max_shift = max_shift
-        
-    def __call__(self, sample):
-        origin, (target , mask) = sample
-        tl_shift = np.random.randint(0, self.max_shift)
-        br_shift = np.random.randint(0, self.max_shift)
-
-        origin_w, origin_h = torchvision.transforms.functional.to_pil_image(origin).size
-        crop_w = origin_w - tl_shift - br_shift
-        crop_h = origin_h - tl_shift - br_shift
-        
-        origin = torchvision.transforms.functional.crop(origin, tl_shift, tl_shift,
-                                                        crop_h, crop_w)
-        mask = torchvision.transforms.functional.crop(mask, tl_shift, tl_shift,
-                                                        crop_h, crop_w)
-        return origin, (target, mask)
-
-class Pad():
-    def __init__(self, max_padding):
-        self.max_padding = max_padding
-        
-    def __call__(self, sample):
-        origin, (target , mask) = sample
-        padding = np.random.randint(0, self.max_padding)
-
-        origin = torchvision.transforms.functional.pad(origin, padding=padding, fill=0)
-        mask = torchvision.transforms.functional.pad(mask, padding=padding, fill=0)
-        return origin, (target, mask)
-
-class Rotate():
-    def __init__(self, max_angle):
-        self.max_angle = max_angle
-        
-    def __call__(self, sample):
-        origin, (target , mask) = sample
-        angle = np.random.randint(-self.max_angle, self.max_angle)
-        
-        origin = torchvision.transforms.functional.rotate(origin, angle)
-        mask = torchvision.transforms.functional.rotate(mask, angle)
-        return origin, (target, mask)
-
-
-class ImageTransform:
-    def __init__(self ):
-        self.image_transform = torchvision.transforms.Compose([
-            Pad(200),
-            Crop(100),
-            Rotate(15)
-        ])
-
-    def __call__(self, sample):
-        return self.image_transform(sample)
-
-class DataAugmentation:
-    def __init__(self, num_samples = 0):
-        self.num_samples = num_samples
-        self.transform = ImageTransform()
-
-    def __call__(self, samples):
-      new_dataset = []
-      for s in samples:
-        for n in range(self.num_samples):
-          new_dataset.append(self.transform(s))
-      samples = samples + new_dataset
-      return samples
